+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Added `Athlete#name` and `Athlete#strava_url` - [@dblock](https://github.com/dblock).
 * [#5](https://github.com/dblock/strava-ruby-client/issues/5): Added support for pagination - [@dblock](https://github.com/dblock).
 * Added `Strava::Api::Client#athlete_clubs` - [@dblock](https://github.com/dblock).
+* Added `Strava::Api::Client#club_activities` - [@dblock](https://github.com/dblock).
 * Your contribution here.
 
 ### 0.1.0 (2018/11/23)